pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gqlcli"
authors = [
    {name = "syfun", email = "sunyu418@gmail.com"},
]
readme = "README.md"
dynamic = ["version", "description"]
requires-python = ">=3.8"
dependencies = [
  "pyclip ~=0.6.0",
  "prompt_toolkit ~=3.0.22",
  "requests ~=2.25.1",
  "graphql-core >=3",
  "click ~=8.0",
]

[project.urls]
Documentation = "https://github.com/syfun/gqlcli"
Source = "https://github.com/syfun/gqlcli"

[project.optional-dependencies]
dev = [
  "black ==22.3.0",
  "flake8 ==4.0.1",
  "isort ==5.10.1",
  "pre-commit ==2.19.0",
]

[project.scripts]
gqlcli = 'gqlcli.main:main'


[tool.isort]
skip_gitignore = true
profile = "black"
line_length = 120
multi_line_output = 3

[tool.black]
line-length = 120
skip-string-normalization = false
exclude = '''
(
  /(
      \.git
    | \.venv
    | \.mypy_cache
    | __pycache__
  )
)
'''
